Rename service-naming/extra as service-naming/extra-services
